Update project setup with Vite and add Github Pages deployment
This commit updates the project setup with Vite and TypeScript. It also adds a new Github Pages deployment workflow. Styles are now managed by postcss and autoprefixer, and the project now uses PNPM instead of NPM. A new dependency, date-fns, was added as well. The previous output CSS file has been deleted as part of this update.