Add `WeeksInfo` type and optimize element class addition
A new type `WeeksInfo` has been added at the top of the file for coherence and better visibility. Additionally, the code where class names are added to `weekDiv` element has been optimized to make it more readable and concise. The `WeeksInfo` type has been deleted from a lower position in the code to avoid repetition.